+++ src/test/java/roomescape/payment/infrastructure/TossPaymentClientTest.java
@@ -8,11 +8,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.test.web.client.ExpectedCount;
 import org.springframework.test.web.client.MockRestServiceServer;
 import org.springframework.web.client.RestClient;
-import roomescape.reservation.exception.PaymentRequestException;
-import roomescape.reservation.exception.RequestPaymentErrorHandler;
+import roomescape.payment.exception.PaymentRequestException;
+import roomescape.payment.exception.RequestPaymentErrorHandler;
 import roomescape.reservation.presentation.dto.PaymentRequest;
